+++ test/behaviour/Gravity.spec.js
@@ -1,6 +1,6 @@
 /*global describe, it */
 
-import * as Proton from '../../src';
+import * as Nebula from '../../src';
 
 import { TIME } from '../constants';
 import chai from 'chai';
@@ -9,7 +9,7 @@ import { getEasingByName } from '../../src/ease';
 const { assert } = chai;
 
 describe('behaviour -> Gravity', () => {
-  const behaviour = new Proton.Gravity(1);
+  const behaviour = new Nebula.Gravity(1);
 
   it('should instantiate with the correct properties', done => {
     const { life, easing, age, energy, dead, force } = behaviour;
@@ -20,20 +20,20 @@ describe('behaviour -> Gravity', () => {
     assert.strictEqual(age, 0);
     assert.strictEqual(energy, 1);
     assert.isFalse(dead);
-    assert.instanceOf(force, Proton.Vector3D);
+    assert.instanceOf(force, Nebula.Vector3D);
     assert.strictEqual(force.y, -100);
 
     done();
   });
 
   it('should have set the correct properties on the particle after applying the behaviour', done => {
-    const particle = new Proton.Particle();
+    const particle = new Nebula.Particle();
 
     behaviour.applyBehaviour(particle, TIME);
 
     const { acceleration, old } = particle;
 
-    assert.instanceOf(acceleration, Proton.Vector3D);
+    assert.instanceOf(acceleration, Nebula.Vector3D);
     assert.strictEqual(acceleration.x, 0);
     assert.strictEqual(acceleration.y, -100);
     assert.strictEqual(acceleration.z, 0);
@@ -45,14 +45,14 @@ describe('behaviour -> Gravity', () => {
   });
 
   it('should construct the behaviour from a JSON object', done => {
-    const instance = Proton.Gravity.fromJSON({
+    const instance = Nebula.Gravity.fromJSON({
       gravity: 1,
       life: 3,
       easing: 'easeInOutExpo',
     });
 
-    assert.instanceOf(instance, Proton.Gravity);
-    assert.instanceOf(instance.force, Proton.Vector3D);
+    assert.instanceOf(instance, Nebula.Gravity);
+    assert.instanceOf(instance.force, Nebula.Vector3D);
     assert.strictEqual(instance.force.x, 0);
     assert.strictEqual(instance.force.y, -100);
     assert.strictEqual(instance.force.z, 0);