+++ test/emitter/Emitter.emitterBehaviours.spec.js
@@ -1,6 +1,6 @@
 /*global describe, it */
 
-import * as Proton from '../../src';
+import * as Nebula from '../../src';
 
 import { TIME } from '../constants';
 import chai from 'chai';
@@ -8,12 +8,12 @@ import sinon from 'sinon';
 
 const { spy } = sinon;
 const { assert } = chai;
-const { Emitter } = Proton;
+const { Emitter } = Nebula;
 
 describe('emitter -> Emitter -> emitterBehaviours', () => {
   it("should add a behaviour to the emitter's behaviours and call the behaviour's initialize method passing the emitter", done => {
     const emitter = new Emitter();
-    const attraction = new Proton.Attraction();
+    const attraction = new Nebula.Attraction();
     const attractionInitializeSpy = spy(attraction, 'initialize');
 
     assert.instanceOf(emitter.addEmitterBehaviour(attraction), Emitter);
@@ -28,25 +28,25 @@ describe('emitter -> Emitter -> emitterBehaviours', () => {
 
   it("should add all the behaviours to the emitter's emitter behaviours", done => {
     const emitter = new Emitter();
-    const attraction = new Proton.Attraction();
-    const repulsion = new Proton.Repulsion();
-    const gravity = new Proton.Gravity();
+    const attraction = new Nebula.Attraction();
+    const repulsion = new Nebula.Repulsion();
+    const gravity = new Nebula.Gravity();
     const behaviours = [attraction, repulsion, gravity];
 
     assert.instanceOf(emitter.addEmitterBehaviours(behaviours), Emitter);
     assert.lengthOf(emitter.emitterBehaviours, behaviours.length);
-    assert.instanceOf(emitter.emitterBehaviours[0], Proton.Gravity);
-    assert.instanceOf(emitter.emitterBehaviours[1], Proton.Repulsion);
-    assert.instanceOf(emitter.emitterBehaviours[2], Proton.Attraction);
+    assert.instanceOf(emitter.emitterBehaviours[0], Nebula.Gravity);
+    assert.instanceOf(emitter.emitterBehaviours[1], Nebula.Repulsion);
+    assert.instanceOf(emitter.emitterBehaviours[2], Nebula.Attraction);
 
     done();
   });
 
   it("should set the emitter's emitter behaviours to the behaviours passed", done => {
     const emitter = new Emitter();
-    const attraction = new Proton.Attraction();
-    const repulsion = new Proton.Repulsion();
-    const gravity = new Proton.Gravity();
+    const attraction = new Nebula.Attraction();
+    const repulsion = new Nebula.Repulsion();
+    const gravity = new Nebula.Gravity();
     const behaviours = [attraction, repulsion, gravity];
 
     assert.instanceOf(emitter.setEmitterBehaviours(behaviours), Emitter);
@@ -58,9 +58,9 @@ describe('emitter -> Emitter -> emitterBehaviours', () => {
 
   it("should remove the emitter's emitter behaviour", done => {
     const emitter = new Emitter();
-    const attraction = new Proton.Attraction();
-    const repulsion = new Proton.Repulsion();
-    const gravity = new Proton.Gravity();
+    const attraction = new Nebula.Attraction();
+    const repulsion = new Nebula.Repulsion();
+    const gravity = new Nebula.Gravity();
     const behaviours = [attraction, repulsion, gravity];
 
     emitter.setEmitterBehaviours(behaviours);
@@ -76,9 +76,9 @@ describe('emitter -> Emitter -> emitterBehaviours', () => {
     const emitter = new Emitter();
 
     emitter.setEmitterBehaviours([
-      new Proton.Attraction(),
-      new Proton.Repulsion(),
-      new Proton.Gravity(),
+      new Nebula.Attraction(),
+      new Nebula.Repulsion(),
+      new Nebula.Gravity(),
     ]);
 
     assert.instanceOf(emitter.removeAllEmitterBehaviours(), Emitter);
@@ -104,11 +104,11 @@ describe('emitter -> Emitter -> updateEmitterBehaviours', () => {
 
   it("should update the emitter's properties after an emitter behaviour has been added and the emitter has been updated", done => {
     const emitter = new Emitter();
-    const attraction = new Proton.Attraction();
-    const repulsion = new Proton.Repulsion();
-    const gravity = new Proton.Gravity();
+    const attraction = new Nebula.Attraction();
+    const repulsion = new Nebula.Repulsion();
+    const gravity = new Nebula.Gravity();
     const behaviours = [attraction, repulsion, gravity];
-    const emitterBehaviour = new Proton.Rotate(1, 0, 0);
+    const emitterBehaviour = new Nebula.Rotate(1, 0, 0);
     const before = { ...emitter.rotation };
     let after;