+++ test/initializer/Life.spec.js
@@ -1,6 +1,6 @@
 /*global describe, it */
 
-import * as Proton from '../../src';
+import * as Nebula from '../../src';
 
 import chai from 'chai';
 
@@ -9,7 +9,7 @@ const { assert } = chai;
 describe('initializer -> Life', () => {
   const start = 3;
   const end = 5;
-  const initializer = new Proton.Life(start, end);
+  const initializer = new Nebula.Life(start, end);
 
   it('should have the correct properties after instantiation', done => {
     const {
@@ -18,7 +18,7 @@ describe('initializer -> Life', () => {
     } = initializer;
 
     assert.equal(initializer.type, 'Life');
-    assert.instanceOf(lifePan, Proton.Span);
+    assert.instanceOf(lifePan, Nebula.Span);
     assert.isFalse(_isArray);
     assert.isFalse(_center);
     assert.strictEqual(a, start);
@@ -28,7 +28,7 @@ describe('initializer -> Life', () => {
   });
 
   it('should set the correct properties on the particle after initialization', done => {
-    const particle = new Proton.Particle();
+    const particle = new Nebula.Particle();
 
     initializer.initialize(particle);
 
@@ -41,8 +41,8 @@ describe('initializer -> Life', () => {
   });
 
   it('should set the particle life to Infinity if the first argument to the constructor is Infinity', done => {
-    const infiniteLife = new Proton.Life(Infinity);
-    const particle = new Proton.Particle();
+    const infiniteLife = new Nebula.Life(Infinity);
+    const particle = new Nebula.Particle();
 
     infiniteLife.initialize(particle);
 
@@ -52,14 +52,14 @@ describe('initializer -> Life', () => {
   });
 
   it('should construct the initializer from a JSON object', done => {
-    const instance = Proton.Life.fromJSON({
+    const instance = Nebula.Life.fromJSON({
       min: 3,
       max: 10,
       center: true,
     });
 
-    assert.instanceOf(instance, Proton.Life);
-    assert.instanceOf(instance.lifePan, Proton.Span);
+    assert.instanceOf(instance, Nebula.Life);
+    assert.instanceOf(instance.lifePan, Nebula.Span);
     assert.equal(instance.lifePan.a, 3);
     assert.equal(instance.lifePan.b, 10);
     assert.isTrue(instance.isEnabled);