+++ test/initializer/Velocity.spec.js
@@ -1,6 +1,6 @@
 /*global describe, it */
 
-import * as Proton from '../../src';
+import * as Nebula from '../../src';
 
 import Velocity from '../../src/initializer/Velocity/Velocity';
 import chai from 'chai';
@@ -11,7 +11,7 @@ describe('initializer -> Velocity abstract', () => {
   const velocity = new Velocity();
 
   it('should set the dirVec property', done => {
-    assert.instanceOf(velocity.dirVec, Proton.Vector3D);
+    assert.instanceOf(velocity.dirVec, Nebula.Vector3D);
 
     done();
   });