+++ webpack.config.js
@@ -2,12 +2,12 @@
 
 const path = require('path');
 
-let libraryName = 'Proton';
+let libraryName = 'Nebula';
 
 let plugins = [],
   outputFile;
 
-outputFile = 'three-proton.js';
+outputFile = 'nebula-engine.js';
 
 const config = {
   entry: __dirname + '/src/index.js',
@@ -17,27 +17,27 @@ const config = {
     filename: outputFile,
     library: libraryName,
     libraryTarget: 'umd',
-    umdNamedDefine: true
+    umdNamedDefine: true,
   },
   module: {
     rules: [
       {
         test: /(\.jsx|\.js)$/,
         loader: 'babel-loader',
-        exclude: /(node_modules|bower_components)/
+        exclude: /(node_modules|bower_components)/,
       },
       {
         test: /(\.jsx|\.js)$/,
         loader: 'eslint-loader',
-        exclude: /node_modules/
-      }
-    ]
+        exclude: /node_modules/,
+      },
+    ],
   },
   resolve: {
     modules: [path.resolve('./src'), path.resolve('./node_modules')],
-    extensions: ['.json', '.js']
+    extensions: ['.json', '.js'],
   },
-  plugins: plugins
+  plugins: plugins,
 };
 
 module.exports = config;